Rename and move MPA RTC linear segment to match current DNC implementation naming and location
